1 拆分单词表与统计表 2 可以通过参数选择重置数据库所有表（-i），以及仅更新单词表，保留统计表（-r），默认无参数则是合并文件词库到db中 3 添加“大家学标准日本语初级”单词表

+++ createdb.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python3
-import os
+import os, getopt
 import sys
 import sqlite3
 import logging
@@ -10,17 +10,16 @@ import Newstandard
 import Welearn
 import Custom
 from Jpdict import Jpdict
+from Jpword import Jpword
 from datetime import timedelta, datetime
 
 dbfile = global_var.get_value('dbfile')
 wordtable = global_var.get_value('wordtable')
+stats = global_var.get_value('stats')
 words = {}
 logger = None
 
-
-
-
-def createdb(items):
+def createtables():
     conn = sqlite3.connect(dbfile)
     cs = conn.cursor()
     sql = '''
@@ -31,48 +30,86 @@ def createdb(items):
         roma TEXT NOT NULL,
         chinese TEXT NOT NULL,
         english TEXT DEFAULT "",
-        
         wordtype TEXT DEFAULT "",
         tone TEXT DEFAULT "",
         lesson TEXT DEFAULT "",
         description TEXT DEFAULT "",
-        nlevel INTEGER DEFAULT 0,
-        
+        nlevel INTEGER DEFAULT 0
+        );
+    '''.format(wordtable)
+    logger.debug(sql)
+    cs.execute(sql)
+
+    sql = '''
+    CREATE TABLE if not exists {0} (
+        kana TEXT PRIMARY KEY,
         increase INTEGER NOT NULL DEFAULT 0,
         decrease INTEGER NOT NULL DEFAULT 0,
         correct INTEGER NOT NULL DEFAULT 0,
         wrong INTEGER NOT NULL DEFAULT 0,
         lasttime REAL NOT NULL DEFAULT 0
         );
-    '''.format(wordtable)
+    '''.format(stats)
     logger.debug(sql)
     cs.execute(sql)
     
-    vals = []
-    ts = datetime.today().timestamp()
-    for kana, value in items:
-        v = [
-            value.kana,
-            value.kanji,
-            value.tone,
-            value.wordtype,
-            value.chinese,
-            ';'.join(value.lesson),
-            value.roma,
-            value.english,
-            ts
-        ]
-
-        vals.append(v)
-    
-    cs.executemany('''
-        INSERT INTO jpwords(kana, kanji, tone, wordtype, chinese, lesson, roma, english, lasttime) VALUES(?,?,?,?,?,?,?,?,?)
-        ''', vals)
-    
-    logger.debug('add {} words'.format(len(items)))
     cs.close()
     conn.commit()
     conn.close()
+
+def getdbwords(jpdict):
+    conn = sqlite3.connect(dbfile)
+    cs = conn.cursor()
+    cs.execute("select * from {}".format(wordtable))
+    data = cs.fetchall()
+    cs.close()
+    conn.close()
+    
+    for tup in data:
+        wordid, kana, kanji, roma, chinese, english, wordtype, tone, lesson, description, nlevel = tup
+        lessonarr = lesson.split(',')
+        jpdict.merge(Jpword(kana, kanji, roma, chinese, english, wordtype, tone, lessonarr, description, nlevel))
+
+
+def resetwordtable(jpdict):
+    conn = sqlite3.connect(dbfile)
+    cs = conn.cursor()
+
+    # 清表
+    sql = 'delete from {}'.format(wordtable)
+    logger.debug(sql)
+    cs.execute(sql)
+    
+    # 自增归0
+    sql = 'update sqlite_sequence SET seq = 0 where name = \'{}\''.format(wordtable)
+    logger.debug(sql)
+    cs.execute(sql)
+
+    # 重新添加所有
+    items = jpdict.getitems()
+    if len(items):
+        vals = []
+        for kana, value in items:
+            v = [
+                value.kana,
+                value.kanji,
+                value.tone,
+                value.wordtype,
+                value.chinese,
+                ';'.join(value.lesson),
+                value.roma,
+                value.english,
+            ]
+            vals.append(v)
+        cs.executemany('''
+            INSERT INTO jpwords(kana, kanji, tone, wordtype, chinese, lesson, roma, english) VALUES(?,?,?,?,?,?,?,?)
+            ''', vals)
+        logger.debug('Total Add {} words to db'.format(len(vals)))
+    cs.close()
+    conn.commit()
+    conn.close()
+
+
 
 if __name__ == '__main__':
     # init log
@@ -81,19 +118,37 @@ if __name__ == '__main__':
         os.unlink(logfile)
     logging.basicConfig(filename = logfile, level = logging.DEBUG, format = '%(asctime)s - %(levelname)s - %(message)s')
     logger = logging
-    
-    # load data
     jpdict = Jpdict()
     
+    try:
+        opts, args = getopt.getopt(sys.argv[1:], '-i-r', ['init', 'reset'])
+    except getopt.GetoptError:
+        print('python3 {}'.format(sys.argv[0]))
+        sys.exit(2)
+    for opt, arg in opts:
+        if opt in ('-i','--init'):
+            # remove db file
+            if os.path.exists(dbfile):
+                os.unlink(dbfile)
+            createtables()
+        elif opt in ('-r','--reset'):
+            # clear word table, Reserve stats table
+            resetwordtable(jpdict)
+            
+    # load word from db
+    getdbwords(jpdict)
+    logger.debug("--- After reading database ---")
+    jpdict.total(logger)
+
+    
+    # load word from file
     # Minnano.load(jpdict, logger)
     # Newstandard.load(jpdict, logger)
     # Custom.load(jpdict, logger)
     Welearn.load(jpdict, logger)
-    
-    # init db
-    if os.path.exists(dbfile):
-        os.unlink(dbfile)
-    createdb(jpdict.getitems())
-    
-    print("add: ", jpdict.totaladd)
-    print("merge: ", jpdict.totalmerge)
+    logger.debug("--- After reading file ---")
+    jpdict.total(logger)
+
+    # reset table
+    resetwordtable(jpdict)
+
